balloon_strategy: correct basis of speed vector to search direction
Previously the speed vector was based on the last spotted direction to
the balloon but this lead to an over correction.  With this change it's
based upon the original yaw and pitch of the balloon when we first
spotted it during the search.
Also added acceleration limiting to the speed vector and a
VEL_DIST_RATIO parameter which allows controlling the target speed based
upon the distance to the balloon.